support for custom option parser in the configglue call
enable support for passing in a custom option parser (instead of just being
able to set the usage text).

config validation depends on the validate parameter and the presence (and
value) of a validate option if a custom option parser is passed in.

by default validation is now turned off